# openapi.json
{"definitions":{"Book":{"properties":{"author":{"description":"Book author","type":"string"},"created_at":{"description":"Book posting timestamp","format":"date-time","type":"string"},"description":{"description":"Book description","type":"string"},"genre":{"description":"Book genre","type":"string"},"id":{"description":"Book ID","type":"integer"},"is_available":{"description":"Whether book is available for taking","type":"boolean"},"meeting_address":{"description":"Pickup/meeting address","type":"string"},"owner_id":{"description":"ID of user who posted the book","type":"integer"},"publish_year":{"description":"Publication year","type":"integer"},"taken_at":{"description":"Timestamp when book was taken","format":"date-time","type":"string"},"taken_by":{"description":"ID of user who took the book (null if available)","type":"integer"},"title":{"description":"Book title","type":"string"},"updated_at":{"description":"Last update timestamp","format":"date-time","type":"string"}},"required":["id","title","author","genre","publish_year","meeting_address","is_available","owner_id","created_at"],"type":"object"},"BookCreateRequest":{"properties":{"author":{"description":"Book author","maxLength":255,"type":"string"},"description":{"description":"Book description","maxLength":255,"type":"string"},"genre":{"description":"Book genre","maxLength":255,"type":"string"},"meeting_address":{"description":"Pickup/meeting address","maxLength":255,"type":"string"},"publish_year":{"description":"Publication year","minimum":0,"type":"integer"},"title":{"description":"Book title","maxLength":255,"type":"string"}},"required":["title","author","genre","publish_year","meeting_address"],"type":"object"},"BookList":{"properties":{"books":{"description":"List of books","items":{"$ref":"#/definitions/Book"},"type":"array"},"total":{"description":"Total number of books matching the criteria","type":"integer"}},"required":["books","total"],"type":"object"},"Error":{"properties":{"error":{"description":"Error message","type":"string"}},"required":["error"],"type":"object"},"LoginRequest":{"properties":{"password":{"description":"Password","type":"string"},"username":{"description":"Username","type":"string"}},"required":["username","password"],"type":"object"},"LoginResponse":{"properties":{"access_token":{"description":"JWT access token","type":"string"}},"required":["access_token"],"type":"object"},"User":{"properties":{"created_at":{"description":"Account creation timestamp","format":"date-time","type":"string"},"id":{"description":"User ID","type":"integer"},"role":{"description":"User role","enum":["user","admin"],"type":"string"},"username":{"description":"Username","type":"string"}},"required":["id","username","role","created_at"],"type":"object"},"UserList":{"properties":{"total":{"description":"Total number of users","type":"integer"},"users":{"description":"List of users","items":{"$ref":"#/definitions/User"},"type":"array"}},"required":["users","total"],"type":"object"},"UserRoleUpdateRequest":{"properties":{"role":{"description":"New user role","enum":["user","admin"],"type":"string"}},"required":["role"],"type":"object"}},"info":{"description":"API for a book sharing platform where users can post books for sharing and take books from others","title":"BookCrossing API","version":"1.0.0"},"paths":{"/api/admin/statistics":{"get":{"responses":{"200":{"description":"Platform statistics","schema":{"properties":{"available_books":{"type":"integer"},"books_created_today":{"type":"integer"},"most_popular_genre":{"type":"string"},"total_books":{"type":"integer"},"total_exchanges":{"type":"integer"},"total_users":{"type":"integer"}},"type":"object"}},"403":{"description":"Admin privileges required"}},"security":[{"Bearer":[]}],"summary":"Get platform statistics (admin only).","tags":["Admin"]}},"/api/admin/users":{"get":{"parameters":[{"default":10,"description":"Number of users to return (1-100)","in":"query","name":"limit","type":"integer"},{"default":0,"description":"Number of users to skip","in":"query","name":"offset","type":"integer"}],"responses":{"200":{"description":"List of users"},"403":{"description":"Admin privileges required"}},"security":[{"Bearer":[]}],"summary":"Get all users (admin only).","tags":["Admin"]}},"/api/admin/users/{user_id}":{"delete":{"parameters":[{"description":"User ID","in":"path","name":"user_id","required":true,"type":"integer"}],"responses":{"200":{"description":"User deleted successfully"},"400":{"description":"Cannot delete yourself"},"403":{"description":"Admin privileges required"},"404":{"description":"User not found"}},"security":[{"Bearer":[]}],"summary":"Delete a user and their books (admin only).","tags":["Admin"]}},"/api/admin/users/{user_id}/role":{"put":{"parameters":[{"description":"User ID","in":"path","name":"user_id","required":true,"type":"integer"},{"in":"body","name":"role_data","required":true,"schema":{"properties":{"role":{"description":"New role for the user","enum":["user","admin"],"type":"string"}},"required":["role"],"type":"object"}}],"responses":{"200":{"description":"User role updated successfully"},"400":{"description":"Invalid role or cannot change your own role"},"403":{"description":"Admin privileges required"},"404":{"description":"User not found"}},"security":[{"Bearer":[]}],"summary":"Change a user's role (admin only).","tags":["Admin"]}},"/api/auth":{"post":{"parameters":[{"description":"User login credentials","in":"body","name":"credentials","required":true,"schema":{"properties":{"password":{"description":"Password","example":"secure_password123","type":"string"},"username":{"description":"Username","example":"john_doe","type":"string"}},"required":["username","password"],"type":"object"}}],"responses":{"200":{"description":"Login successful","schema":{"properties":{"access_token":{"description":"JWT access token","type":"string"}},"type":"object"}},"401":{"description":"Invalid credentials","schema":{"properties":{"error":{"type":"string"}},"type":"object"}}},"summary":"Authenticate user and return JWT token.","tags":["Authentication"]}},"/api/auth/me":{"get":{"responses":{"200":{"description":"Current user information","schema":{"properties":{"user":{"properties":{"created_at":{"format":"date-time","type":"string"},"id":{"type":"integer"},"role":{"type":"string"},"username":{"type":"string"}},"type":"object"}},"type":"object"}},"401":{"description":"Authentication required","schema":{"properties":{"error":{"type":"string"}},"type":"object"}}},"security":[{"Bearer":[]}],"summary":"Get current authenticated user information.","tags":["Authentication"]}},"/api/auth/register":{"post":{"parameters":[{"description":"User registration data","in":"body","name":"user","required":true,"schema":{"properties":{"password":{"description":"User password","example":"secure_password123","type":"string"},"username":{"description":"Unique username (max 255 characters)","example":"john_doe","type":"string"}},"required":["username","password"],"type":"object"}}],"responses":{"201":{"description":"User successfully registered","schema":{"properties":{"message":{"example":"User registered successfully","type":"string"},"user":{"properties":{"id":{"type":"integer"},"role":{"type":"string"},"username":{"type":"string"}},"type":"object"}},"type":"object"}},"400":{"description":"Invalid request data or username already exists","schema":{"properties":{"error":{"type":"string"}},"type":"object"}}},"summary":"Register a new user account.","tags":["Authentication"]}},"/api/books":{"get":{"parameters":[{"default":10,"description":"Number of books to return (1-100)","in":"query","name":"limit","type":"integer"},{"default":0,"description":"Number of books to skip","in":"query","name":"offset","type":"integer"},{"description":"Filter by book title (partial match)","in":"query","name":"title","type":"string"},{"description":"Filter by author name (partial match)","in":"query","name":"author","type":"string"},{"description":"Filter by genre (partial match)","in":"query","name":"genre","type":"string"},{"description":"Show only available books","in":"query","name":"available_only","type":"boolean"},{"description":"Filter by publication year","in":"query","name":"publish_year","type":"integer"}],"responses":{"200":{"description":"List of books","schema":{"properties":{"books":{"items":{"properties":{"author":{"type":"string"},"genre":{"type":"string"},"id":{"type":"integer"},"is_available":{"type":"boolean"},"publish_year":{"type":"integer"},"title":{"type":"string"}},"type":"object"},"type":"array"},"limit":{"type":"integer"},"offset":{"type":"integer"},"total":{"description":"Total number of books matching filters","type":"integer"}},"type":"object"}}},"security":[{"Bearer":[]}],"summary":"Get books with pagination and optional filters.","tags":["Books"]},"post":{"parameters":[{"description":"Book data","in":"body","name":"book","required":true,"schema":{"properties":{"author":{"description":"Book author (max 255 characters)","example":"F. Scott Fitzgerald","type":"string"},"description":{"description":"Optional book description (max 255 characters)","example":"Great condition, minimal wear","type":"string"},"genre":{"description":"Book genre (max 255 characters)","example":"Classic Literature","type":"string"},"meeting_address":{"description":"Pickup address (max 255 characters)","example":"123 Main St, City, State","type":"string"},"publish_year":{"description":"Publication year","example":1925,"type":"integer"},"title":{"description":"Book title (max 255 characters)","example":"The Great Gatsby","type":"string"}},"required":["title","author","publish_year","genre","meeting_address"],"type":"object"}}],"responses":{"201":{"description":"Book created successfully","schema":{"properties":{"book":{"type":"object"},"message":{"type":"string"}},"type":"object"}},"400":{"description":"Invalid request data"}},"security":[{"Bearer":[]}],"summary":"Create a new book posting.","tags":["Books"]}},"/api/books/my":{"get":{"parameters":[{"default":10,"description":"Number of books to return (1-100)","in":"query","name":"limit","type":"integer"},{"default":0,"description":"Number of books to skip","in":"query","name":"offset","type":"integer"}],"responses":{"200":{"description":"List of user's books","schema":{"properties":{"books":{"items":{"type":"object"},"type":"array"},"limit":{"type":"integer"},"offset":{"type":"integer"},"total":{"type":"integer"}},"type":"object"}}},"security":[{"Bearer":[]}],"summary":"Get current user's posted books.","tags":["Books"]}},"/api/books/taken":{"get":{"parameters":[{"default":10,"description":"Number of books to return (1-100)","in":"query","name":"limit","type":"integer"},{"default":0,"description":"Number of books to skip","in":"query","name":"offset","type":"integer"}],"responses":{"200":{"description":"List of books taken by user","schema":{"properties":{"books":{"items":{"type":"object"},"type":"array"},"limit":{"type":"integer"},"offset":{"type":"integer"},"total":{"type":"integer"}},"type":"object"}}},"security":[{"Bearer":[]}],"summary":"Get books taken by current user.","tags":["Books"]}},"/api/books/{book_id}":{"delete":{"parameters":[{"description":"Book ID","in":"path","name":"book_id","required":true,"type":"integer"}],"responses":{"200":{"description":"Book deleted successfully"},"403":{"description":"Cannot delete this book"},"404":{"description":"Book not found"}},"security":[{"Bearer":[]}],"summary":"Delete a book posting.","tags":["Books"]},"put":{"parameters":[{"description":"Book ID","in":"path","name":"book_id","required":true,"type":"integer"},{"description":"Updated book data","in":"body","name":"book","required":true,"schema":{"properties":{"author":{"type":"string"},"description":{"type":"string"},"genre":{"type":"string"},"meeting_address":{"type":"string"},"publish_year":{"type":"integer"},"title":{"type":"string"}},"required":["title","author","publish_year","genre","meeting_address"],"type":"object"}}],"responses":{"200":{"description":"Book updated successfully"},"400":{"description":"Invalid request data"},"403":{"description":"Cannot modify this book"},"404":{"description":"Book not found"}},"security":[{"Bearer":[]}],"summary":"Update a book posting.","tags":["Books"]}},"/api/books/{book_id}/take":{"post":{"parameters":[{"description":"Book ID","in":"path","name":"book_id","required":true,"type":"integer"}],"responses":{"200":{"description":"Book taken successfully"},"400":{"description":"Cannot take this book"},"404":{"description":"Book not found"}},"security":[{"Bearer":[]}],"summary":"Take a book from another user.","tags":["Books"]}},"/health":{"get":{"responses":{"200":{"description":"Application is healthy","schema":{"properties":{"status":{"example":"healthy","type":"string"},"timestamp":{"example":"2025-08-29T09:52:00Z","type":"string"}},"type":"object"}}},"summary":"Health check endpoint for monitoring and load balancers.","tags":["Health"]}}},"security":[{"Bearer":[]}],"securityDefinitions":{"Bearer":{"description":"JWT authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token.","in":"header","name":"Authorization","type":"apiKey"}},"swagger":"2.0"}